\name{epiplus-package}


\alias{epiplus-package}
\alias{epiplus}

\docType{package}

\title{
  epiplus
}

\description{
Complementary epidemiology indicators and measures.
}

\details{
  \tabular{ll}{
    Package: \tab  epiplus \cr
    Type: \tab
    Package\cr
    Version: \tab  0.1.0 \cr
    Date: \tab  01/10/2022 \cr
    License: \tab neant \cr
    LazyLoad: \tab yes\cr
  }


}

\author{
  Hugo Marthinet  <h.marthinet@gmail.com> \cr
}

\references{

}

\seealso{
  Other packages by the same authors :

  \code{\link{lazyMe}}, \code{\link{dataDesc}}, \code{\link{manageR}}, \code{\link{surveyscores}}, \code{\link{sampleSize}}
}

\examples{

}
